Moving sigproc functions to io module. Restructuring params. common Header
