+++ setup.cfg
@@ -30,6 +30,7 @@ install_requires =
     astropy>=4.0
     rich
     attrs
+    bidict
     importlib_metadata; python_version<'3.8'
 
 [options.entry_points]
