+++ sigpyproc/params.py
@@ -1,100 +1,18 @@
-import attr
 import numpy as np
 
-from typing import Optional, ClassVar, Dict, Any, Tuple, Callable
-
-nbits_to_dtype = {1: "<u1", 2: "<u1", 4: "<u1", 8: "<u1", 16: "<u2", 32: "<f4"}
-
-
-@attr.s(auto_attribs=True)
-class BitsInfo(object):
-    """Class to handle bits info.
-
-    Raises
-    ------
-    ValueError
-        if input `nbits` not in [1, 2, 4, 8, 16, 32]
-    """
-
-    nbits: int = attr.ib(validator=attr.validators.in_(nbits_to_dtype.keys()))
-    digi_sigma: float = attr.ib()
-
-    float_bits: ClassVar[int] = 32
-    default_sigma: ClassVar[Dict[int, float]] = {
-        1: 0.5,
-        2: 1.5,
-        4: 6,
-        8: 6,
-        16: 6,
-        32: 6,
-    }
-
-    def __attrs_post_init__(self) -> None:
-        self._digi_min = 0
-        self._digi_max = (1 << self.nbits) - 1
-        self._digi_mean = (1 << (self.nbits - 1)) - 0.5
-        self._digi_scale = self._digi_mean / self.digi_sigma
-
-    @property
-    def dtype(self) -> np.dtype:
-        """Type of the data (`np.dtype`, read-only)."""
-        return np.dtype(nbits_to_dtype[self.nbits])
-
-    @property
-    def itemsize(self) -> int:
-        """Element size of this data-type object (`int`, read-only)."""
-        return self.dtype.itemsize
-
-    @property
-    def unpack(self) -> bool:
-        """Whether to unpack bits (`bool`, read-only)."""
-        return bool(self.nbits in {1, 2, 4})
-
-    @property
-    def bitfact(self) -> int:
-        """Bit factor to unpack/pack bits (`int`, read-only)."""
-        return 8 // self.nbits if self.unpack else 1
-
-    @property
-    def digi_min(self) -> Optional[int]:
-        """Minimum value used to quantize data (`int` or None, read-only)."""
-        return None if self.nbits == self.float_bits else self._digi_min
-
-    @property
-    def digi_max(self) -> Optional[int]:
-        """Maximum value used to quantize data (`int` or None, read-only)."""
-        return None if self.nbits == self.float_bits else self._digi_max
-
-    @property
-    def digi_mean(self) -> Optional[float]:
-        """Mean used to quantize data (`float` or None, read-only)."""
-        return None if self.nbits == self.float_bits else self._digi_mean
-
-    @property
-    def digi_scale(self) -> Optional[float]:
-        """Scale used to quantize data (`float` or None, read-only)."""
-        return None if self.nbits == self.float_bits else self._digi_scale
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Get a dict of all property attributes.
-
-        Returns
-        -------
-        dict
-            property attributes
-        """
-        return {
-            key: getattr(self, key)
-            for key, value in vars(type(self)).items()  # noqa: WPS421
-            if isinstance(value, property)
-        }
-
-    @digi_sigma.default
-    def _set_digi_sigma(self):
-        return self.default_sigma[self.nbits]
+from bidict import bidict
+from astropy import units, constants
+from typing import Dict, Tuple, Callable
 
 
-bits_info = {nbits: BitsInfo(nbits) for nbits in nbits_to_dtype.keys()}
+DM_CONSTANT_LK = 4.148808e3  # L&K Handbook of Pulsar Astronomy
+DM_CONSTANT_MT = 1 / 0.000241  # TEMPO2 Manchester & Taylor (1972)  # noqa: WPS432
+DM_CONSTANT_SI = (
+    (constants.e.esu ** 2 / (2 * np.pi * constants.m_e * constants.c)).to(
+        units.s * units.MHz ** 2 * units.cm ** 3 / units.pc
+    )
+).value  # Precise SI constants
+
 
 # dictionary to define the sizes of header elements
 header_keys = {
@@ -142,28 +60,6 @@ header_keys = {
 }
 
 # header keys recognised by the sigproc package
-sigproc_keys = [
-    "signed",
-    "telescope_id",
-    "ibeam",
-    "nbeams",
-    "refdm",
-    "nifs",
-    "nchans",
-    "foff",
-    "fch1",
-    "nbits",
-    "tsamp",
-    "tstart",
-    "src_dej",
-    "src_raj",
-    "za_start",
-    "az_start",
-    "source_name",
-    "rawdatafile",
-    "data_type",
-    "machine_id",
-]
 
 # header units for fancy printing
 header_units = {
@@ -198,42 +94,22 @@ header_units = {
 }
 
 # data type flag for sigproc files
-data_types = {1: "Filterbank file", 2: "Timeseries file"}
+data_types = bidict(
+    {
+        0: "raw data",
+        1: "filterbank",
+        2: "time series",
+        3: "pulse profiles",
+        4: "amplitude spectrum",
+        5: "complex spectrum",
+        6: "dedispersed subbands",
+        10: "PSRFITS",
+    }
+)
 
 # convert between types from the struct module and numpy
 struct_to_numpy = {"I": "uint", "d": "float", "str": "S256"}
 
-telescope_ids = {
-    "Fake": 0,
-    "Arecibo": 1,
-    "Ooty": 2,
-    "Nancay": 3,
-    "Parkes": 4,
-    "Jodrell": 5,
-    "GBT": 6,
-    "GMRT": 7,
-    "Effelsberg": 8,
-    "Effelsberg LOFAR": 9,
-    "SRT": 10,
-    "Unknown": 11,
-}
-
-ids_to_telescope = dict(zip(telescope_ids.values(), telescope_ids.keys()))
-
-machine_ids = {
-    "FAKE": 0,
-    "PSPM": 1,
-    "Wapp": 2,
-    "AOFTM": 3,
-    "BCPM1": 4,
-    "OOTY": 5,
-    "SCAMP": 6,
-    "GBT Pulsar Spigot": 7,
-    "PFFTS": 8,
-    "Unknown": 9,
-}
-
-ids_to_machine = dict(zip(machine_ids.values(), machine_ids.keys()))
 
 # not required (may be of use in future)
 telescope_lats_longs = {"Effelsberg": (50.52485, 6.883593)}
@@ -242,7 +118,7 @@ telescope_lats_longs = {"Effelsberg": (50.52485, 6.883593)}
 presto_inf: Dict[str, Tuple[str, Callable, str]] = {
     "Data file name without suffix": ("basename", str, "s"),
     "Telescope used": ("telescope", str, "s"),
-    "Instrument used": ("machine", str, "s"),
+    "Instrument used": ("backend", str, "s"),
     "Object being observed": ("source_name", str, "s"),
     "J2000 Right Ascension (hh:mm:ss.ssss)": ("ra", str, "s"),
     "J2000 Declination     (dd:mm:ss.ssss)": ("dec", str, "s"),
